+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added "notification" analytics event [#5611](https://github.com/rokwire/illinois-app/issues/5611).
+
 ## [7.3.38] - 2025-12-23
 ### Changed
 - Building Names now show abbreviated short names too [#5476](https://github.com/rokwire/illinois-app/issues/5476).